+++ visualization/asset_allocation/returns.py
@@ -3,7 +3,9 @@ from matplotlib.ticker import PercentFormatter
 import pandas as pd
 
 
-def strategy_returns(strategy_df: pd.DataFrame) -> None:
+def strategy_returns(
+    strategy_df: pd.DataFrame, excluded_allocations: list = []
+) -> None:
     """
     For a given DataFrame of returns, plot the line chart of cumulative returns
 
@@ -11,6 +13,8 @@ def strategy_returns(strategy_df: pd.DataFrame) -> None:
     ----------
     strategy_df: pd.DataFrame
         DataFrame with portfolio_name and daily returns
+    excluded_allocations: list, optional
+        list of startegies that should not be shown in graph
     """
     strategy_df = strategy_df.dropna()
     strategy_df["return"] = strategy_df["return"] + 1
@@ -19,7 +23,7 @@ def strategy_returns(strategy_df: pd.DataFrame) -> None:
 
     fig, ax = plt.subplots(figsize=(8, 6))
     for name, group in grouped:
-        if name == "ex_ante_mean_variance":
+        if name in excluded_allocations:
             continue
         grouped_df = group["return"].cumprod()
         grouped_df.plot(ax=ax, label=name)