+++ CHANGELOG.md
@@ -8,8 +8,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- Pick-All strategy
+- OLS linear regression calculation 
+- ridge regression calculation 
+- test folder for mathstats
+- overwrite history function for Snowflake
 ### Fixed
+
 ### Changed
+- allow quandl pull without ticker list
+- add MSCI CCC screen for Sovereigns
+
 ### Removed